Moving tests for DB fields into separate 'test_db_fields.py' file.
